StockService: Fixed models, saving enums as strings, matching types in dto Co-authored-by: name <m.ljubic9802@gmail.com>

+++ stock-service/src/main/java/rs/banka4/stock_service/domain/security/future/dtos/FutureDto.java
@@ -21,7 +21,7 @@ public record FutureDto(
     @Schema(
         description = "Size of the future contract",
         example = "1000"
-    ) int contractSize,
+    ) long contractSize,
 
     @Schema(
         description = "Unit of the contract size",
